feat: Add a gatus patch to allow ordering in the UI.
The patch introduces two new configuration keys under endpoint -
group-order and within-group-order.

- group-order controls how the different groups need to be relatively
  ordered (i.e. lexicographically based on the string specified here).
- within-group-order is like group order but for ordering the endpoints
  within the same group based on this string.

CAUTION: Since Gatus uses the endpoint keys as the means to control ordering
in the UI, this patch does change the primary key for the endpoint rows in
the database, and hence will require throwing away the old database
entirely and starting from scratch.

Any endpoints defined without specifying a value for either group-order
or within-group-order default to a value of `z` for both. In other
words, those entries will appear last and continue to get sorted like
they were before in determining the display order for the UI.